Script Usability & Resilience

+++ 07_stop_definode.sh
@@ -5,8 +5,15 @@ docker rm defi_wallet
 docker stop defi_node
 docker rm defi_node
 
-echo definode stopped
-
 ## remove desktop shortcut
 cd /home/pi/Desktop
-rm ./defiWalletShortcut.desktop
+if [ -f "defiWalletShortcut.desktop" ]; then
+	rm ./defiWalletShortcut.desktop
+fi
+
+echo ""
+echo "***************************************************************************"
+echo "DeFi Node has been stopped
+echo "***************************************************************************"
+echo ""
+echo ""
